refactor: DSW-0 updated components to new metadata title format for nextjs 14
