+++ .gitignore
@@ -107,6 +107,9 @@ next-env.d.ts
 .next/
 **/out/
 
+# Next.js old file structure
+nextjs-app/
+
 # Nuxt.js build / generate output
 .nuxt
 dist
