+++ nextjs-app-v14/src/app/components/assistive-text/assistive-text.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import NavigationLayout from "@/app/layout/navigation";
+import { PieAssistiveText } from '@justeattakeaway/pie-webc/react/assistive-text.js';
+
+export default function AssistiveText() {
+    return (
+        <NavigationLayout title="Assistive Text">
+        <PieAssistiveText>Assistive Text</PieAssistiveText>
+        <PieAssistiveText variant="success">Assistive Text</PieAssistiveText>
+        <PieAssistiveText variant="error">Assistive Text</PieAssistiveText>
+        </NavigationLayout>
+    );
+}
