+++ nextjs-app-v14/src/app/components/checkbox/checkbox.tsx
@@ -1,0 +1,23 @@
+'use client';
+
+import NavigationLayout from "@/app/layout/navigation";
+import { PieCheckbox } from "@justeattakeaway/pie-webc/react/checkbox.js";
+import { useState } from "react";
+
+export default function Checkbox() {
+    const [isCheckboxChecked, setIsCheckboxCheck] = useState(false)
+
+    const handleCheckboxChange = () => {
+        setIsCheckboxCheck(current => !current);
+    };
+
+    return (
+        <NavigationLayout title="Checkbox">
+            <PieCheckbox
+                label={`checked: ${isCheckboxChecked}`}
+                checked={isCheckboxChecked}
+                onChange={handleCheckboxChange}>
+            </PieCheckbox>
+        </NavigationLayout>
+    );
+}
