+++ nextjs-app-v14/src/app/components/checkbox/page.tsx
@@ -1,23 +1,10 @@
-'use client';
+import  Checkbox  from './checkbox';
+import { Metadata } from 'next'
 
-import NavigationLayout from "@/layout/navigation";
-import { PieCheckbox } from "@justeattakeaway/pie-webc/react/checkbox.js";
-import { useState } from "react";
+export const metadata: Metadata = {
+    title: 'Checkbox',
+}
 
-export default function Checkbox() {
-    const [isCheckboxChecked, setIsCheckboxCheck] = useState(false)
-
-    const handleCheckboxChange = () => {
-        setIsCheckboxCheck(current => !current);
-    };
-
-    return (
-        <NavigationLayout title="Checkbox">
-            <PieCheckbox
-                label={`checked: ${isCheckboxChecked}`}
-                checked={isCheckboxChecked}
-                onChange={handleCheckboxChange}>
-            </PieCheckbox>
-        </NavigationLayout>
-    );
+export default function CheckboxComponent() {
+    return <Checkbox/>;
 }
