+++ nextjs-app-v14/src/app/components/chip/page.tsx
@@ -1,17 +1,10 @@
-'use client';
+import  Chip  from './chip';
+import { Metadata } from 'next'
 
-import NavigationLayout from "@/layout/navigation";
-import { PieChip } from '@justeattakeaway/pie-webc/react/chip.js';
-import { PieDivider } from '@justeattakeaway/pie-webc/react/divider.js';
+export const metadata: Metadata = {
+    title: 'Chip',
+}
 
-export default function Chip() {
-    return (
-        <NavigationLayout title="Chip">
-        <PieChip>PIE Chip</PieChip>
-        <PieDivider/>
-        <PieChip variant="outline">PIE Chip</PieChip>
-        <PieDivider/>
-        <PieChip variant="ghost">PIE Chip</PieChip>
-        </NavigationLayout>
-    );
+export default function ChipComponent() {
+    return <Chip/>;
 }
