+++ nextjs-app-v14/src/app/components/form-label/form-label.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+import NavigationLayout from "@/app/layout/navigation";
+import { PieFormLabel } from '@justeattakeaway/pie-webc/react/form-label.js';
+
+export default function FormLabel() {
+    return (
+        <NavigationLayout title="Form Label">
+        <PieFormLabel>Label</PieFormLabel>
+        </NavigationLayout>
+    );
+}
