+++ nextjs-app-v14/src/app/components/home-page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import NavigationLayout from "@/layout/navigation";
+import NavigationLayout from "@/app/layout/navigation";
 import { PieLink } from '@justeattakeaway/pie-link/dist/react';
 import { useRouter } from 'next/navigation';
