+++ nextjs-app-v14/src/app/components/link/link.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+import NavigationLayout from "@/app/layout/navigation";
+import { PieLink } from '@justeattakeaway/pie-webc/react/link.js';
+
+export default function Link() {
+    return (
+        <NavigationLayout title="Link">
+        <PieLink href="https://pie.design" target="_blank">Visit pie.design</PieLink>
+        </NavigationLayout>
+    );
+}
