+++ nextjs-app-v14/src/app/components/switch/page.tsx
@@ -1,24 +1,10 @@
-'use client';
+import  Switch  from './switch';
+import { Metadata } from 'next'
 
-import NavigationLayout from "@/layout/navigation";
-import { PieSwitch } from '@justeattakeaway/pie-webc/react/switch.js';
-import { useState } from "react";
+export const metadata: Metadata = {
+    title: 'Switch',
+}
 
-export default function Switch() {
-    const [isSwitchChecked, setIsSwitchCheck] = useState(false)
-
-    const handleSwitchChange = () => {
-      setIsSwitchCheck(current => !current);
-      console.log(isSwitchChecked);
-    };
-
-    return (
-        <NavigationLayout title="Switch">
-        <PieSwitch
-            label={`checked: ${isSwitchChecked}`}
-            checked={isSwitchChecked}
-            onChange={handleSwitchChange}>
-        </PieSwitch>
-        </NavigationLayout>
-    );
+export default function SwitchComponent() {
+    return <Switch/>;
 }
