+++ nextjs-app-v14/src/app/components/tag/tag.tsx
@@ -1,0 +1,12 @@
+'use client';
+
+import NavigationLayout from "@/app/layout/navigation";
+import { PieTag } from '@justeattakeaway/pie-webc/react/tag.js';
+
+export default function Tag() {
+    return (
+        <NavigationLayout title="Tag">
+        <PieTag variant="brand" size="large">Label</PieTag>
+        </NavigationLayout>
+    );
+}
