+++ nextjs-app-v14/src/app/integrations/icons/icons.tsx
@@ -1,0 +1,16 @@
+'use client';
+
+import NavigationLayout from '@/app/layout/navigation';
+import * as icons from '@justeattakeaway/pie-icons-webc/dist/react';
+
+export default function Icons() {
+    return (
+        <NavigationLayout title="Icons Test Page">
+            <div className={"icon-container"}>
+                {Object.entries(icons).map(([iconName, IconComponent]) => (
+                    <IconComponent key={iconName} />
+                ))}
+            </div>
+        </NavigationLayout>
+    );
+}
